Conflicts:
	fabfile.py
	glynt/apps/engage/__init__.py
	glynt/settings.py
	requirements.txt

+++ fabfile.py
@@ -4,6 +4,8 @@ from fabric.contrib.console import confirm
 from fabric.context_managers import settings
 from fabric.contrib import files
 
+from git import *
+
 import os
 import json
 import getpass
@@ -16,10 +18,17 @@ from pprint import pprint
 debug = True
 
 env.local_project_path = os.path.dirname(os.path.realpath(__file__))
+
+env.repo = Repo(env.local_project_path)
+
 env.SHA1_FILENAME = None
 env.timestamp = time.time()
 env.is_predeploy = False
 env.local_user = getpass.getuser()
+env.environment = 'local'
+
+env.truthy = ['true','t','y','yes','1',1]
+env.falsy = ['false','f','n','no','0',0]
 
 @task
 def prod_db():
@@ -190,13 +199,56 @@ def db_backup(db='lawpal_production'):
     local('scp -i %s %s@%s:/tmp/%s /tmp/' % (env.key_filename, env.user, env.host, db_backup_name,))
 
 @task
-def db_local_restore(db='lawpal_production', db_file=None):
+def db_restore(db='lawpal_production', db_file=None):
     with settings(warn_only=True): # only warning as we will often have errors importing
         if db_file is None:
             db_file = '/tmp/%s.bak' % db
-        local('echo "DROP DATABASE %s;" | psql -h localhost -U %s' % (db, env.local_user,))
-        local('echo "CREATE DATABASE %s WITH OWNER %s ENCODING \'UTF8\';" | psql -h localhost -U %s' % (db, env.local_user, env.local_user,))
-        local('pg_restore -U %s -h localhost -d %s -Fc %s' % (env.local_user, db, db_file,))
+            if not os.path.exists(db_file):
+                print colored('Database Backup %s does not exist...' % db_file, 'red')
+            else:
+                go = prompt(colored('Restore "%s" DB from a file entitled: "%s" in the "%s" environment: Proceed? (y,n)' % (db, db_file, env.environment,), 'yellow'))
+                if go in env.truthy:
+                    local('echo "DROP DATABASE %s;" | psql -h localhost -U %s' % (db, env.local_user,))
+                    local('echo "CREATE DATABASE %s WITH OWNER %s ENCODING \'UTF8\';" | psql -h localhost -U %s' % (db, env.local_user, env.local_user,))
+                    local('pg_restore -U %s -h localhost -d %s -Fc %s' % (env.local_user, db, db_file,))
+
+@task
+def git_tags():
+    """ returns list of tags """
+    tags = env.repo.tags
+    return tags
+
+@task
+def git_previous_tag():
+    # last tag in list
+    previous = git_tags()[-1]
+    return previous
+
+@task
+def git_suggest_tag():
+    """ split into parts v1.0.0 drops v converts to ints and increaments and reassembles v1.0.1"""
+    previous = git_previous_tag().name.split('.')
+    mapped = map(int, previous[1:]) # convert all digits to int but exclude the first one as it starts with v and cant be an int
+    next = [int(previous[0].replace('v',''))] + mapped #remove string v and append mapped list
+    next_rev = next[2] = mapped[-1] + 1 # increment the last digit
+    return {
+        'next': 'v%s' % '.'.join(map(str,next)), 
+        'previous': '.'.join(previous)
+    }
+
+@task
+def git_set_tag():
+    proceed = prompt(colored('Do you want to tag this realease?', 'red'), default='y')
+    if proceed in env.truthy:
+        suggested = git_suggest_tag()
+        tag = prompt(colored('Please enter a tag: previous: %s suggested: %s' % (suggested['previous'], suggested['next']), 'yellow'), default=suggested['next'])
+        if tag:
+            tag = 'v%s' % tag if tag[0] != 'v' else tag # ensure we start with a "v"
+
+            message = env.deploy_desc if 'deploy_desc' in env else prompt(colored('Please enter a tag comment', 'green'))
+            env.repo.create_tag(message=message)
+#            local('git tag -a %s -m "%s"' % (tag, comment))
+#            local('git push origin %s' % tag)
 
 @task
 def git_export(branch='aws-sqs'):
@@ -513,14 +565,15 @@ def deploy(is_predeploy='False',full='False',db='False',search='False'):
     :is_predeploy=True - will deploy the latest MASTER SHA but not link it in: this allows for assets collection
     and requirements update etc...
     """
-    true_list = ['true','t','y','yes','1',1]
-    env.is_predeploy = True if is_predeploy.lower() in true_list else False
-    full = True if full.lower() in true_list else False
-    db = True if db.lower() in true_list else False
-    search = True if search.lower() in true_list else False
+    env.is_predeploy = is_predeploy.lower() in env.truthy
+    full = full.lower() in env.truthy
+    db = db.lower() in env.truthy
+    search = search.lower() in env.truthy
 
     diff()
     newrelic_note()
+    git_set_tag()
+
     prepare_deploy()
     do_deploy()
     update_env_conf()
